fix: minor improvements and organize code
