+++ src/translations/English.json
@@ -771,6 +771,7 @@
   "Lightning Node":"Lightning Node",
   "Confirm launch":"Confirm launch",
   "You’re all set!":"You’re all set!",
-  "Docs":"Docs"
+  "Docs":"Docs",
+  "Zoom":"Zoom"
 
 }
